Extract get_from_s3 into library.
This makes it possible to get files from s3 outside the normal dsl.
